# Improve factorize!, type pretty-printing, getindex and fix other issues