--- a/Project.toml
+++ b/Project.toml
@@ -4,9 +4,9 @@ version = "2.0.1"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df" 
-StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
+StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
 Requires = "1"
@@ -14,10 +14,10 @@ StatsBase = "0.33"
 julia = "1.3"
 
 [extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
-PooledArrays = "2dfb63ee-cc39-5dd5-95bd-886bf059d720"
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
+PooledArrays = "2dfb63ee-cc39-5dd5-95bd-886bf059d720"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test", "CUDA", "PooledArrays", "CategoricalArrays"]