I went through a portion of the source code and made the following changes (with some tests included).

### Enhancement

1. Replace `factorize!` with a more efficient implementation (see benchmark results and more details at the bottom).
2. Improve the methods of `show` for `FixedEffect`.
3. Improve `getindex` for `FixedEffect` and add the methods for `GroupedArray`.

### Bug Fixes

1. `DimensionError` does not exist. It should be `DimensionMismatch`.
2. `eltype(fe::FixedEffect) = eltype(I)` is wrong (`I` is something else without the type parameters).
3. `size(g::GroupedArray)` was not defined. This causes error when trying to show `g`.

### Minor Changes

1. The type of the object returned by `get(invpool, x, 0)` can be either `UInt32` or `Int64`. For operations inside such a tight loop, it seems to be slightly better to only have `UInt32`.
2. Remove the redundant `has_missing` inside `group`.
3. Define `size` for `FixedEffect`.
4. Packages listed in `Project.toml` did not follow alphabetical order. `Pkg` will want to correct this order when doing something with the package environment.

## Additional Details

### Changes on `factorize!`

1. There is no need to first find the unique values before relabeling. One loop is enough and the dictionary will handle that, just like how things work with `group`. (`Dict` is fast for telling whether an object has been met before.)
2. I did a rough benchmark with results shown below:
```julia
using BenchmarkTools
using FixedEffects: GroupedArray, group, factorize!
using StatsBase

const n = 10000000
const a = Vector{Union{Float64, Missing}}(undef, n)
sample!(1:50000, a)
a[500:500:n] .= missing

function old_factorize!(x::GroupedArray{N}) where {N}
	refs = x.refs
	uu = sort!(unique(refs))
	has_missing = uu[1] == 0
	ngroups = length(uu) - has_missing
	dict = Dict{UInt32, UInt32}(zip(uu, UInt32(1-has_missing):UInt32(ngroups)))
	for i in eachindex(refs)
		refs[i] = dict[refs[i]]
	end
	GroupedArray{N}(refs, ngroups)
end

@benchmark old_factorize!(g) setup=(g=group(a))
#= Old factorize!
BenchmarkTools.Trial: 
  memory estimate:  5.17 MiB
  allocs estimate:  82
  --------------
  minimum time:     167.921 ms (0.00% GC)
  median time:      174.756 ms (0.00% GC)
  mean time:        174.487 ms (0.11% GC)
  maximum time:     179.650 ms (0.00% GC)
  --------------
  samples:          17
  evals/sample:     1
=#

@benchmark factorize!(g) setup=(g=group(a))
#= New factorize!
BenchmarkTools.Trial: 
  memory estimate:  3.00 MiB
  allocs estimate:  35
  --------------
  minimum time:     79.639 ms (0.00% GC)
  median time:      81.469 ms (0.00% GC)
  mean time:        83.239 ms (0.00% GC)
  maximum time:     93.849 ms (0.00% GC)
  --------------
  samples:          24
  evals/sample:     1
=#

# Compare performance with group (factorize! should be faster than group)
@benchmark group($a)
#=
BenchmarkTools.Trial: 
  memory estimate:  42.82 MiB
  allocs estimate:  38
  --------------
  minimum time:     120.100 ms (0.00% GC)
  median time:      124.897 ms (0.00% GC)
  mean time:        128.562 ms (3.36% GC)
  maximum time:     180.381 ms (31.43% GC)
  --------------
  samples:          39
  evals/sample:     1
=#
```

### Changes on `show(io::IO, fe::FixedEffect)`

1. It is better to only show something short with this method. (It is used when `fe` is an element for some other container.) The additional details can be printed with `show(io::IO, ::MIME"text/plain", fe::FixedEffect)`. This method is used when the result is directly printed to `REPL`.
2. There was an extra newline from the last `println`. That's why `print(io, '\n', ...)` is often used.