@matthieugomez 2021-02-25
Looks great. Can you check that FixedEffectModels and InteractiveFixedEffectModels still pass the test after these changes?