@junyuan-chen 2021-02-25
> Looks great. Thanks a lot. Can you check that FixedEffectModels and InteractiveFixedEffectModels still pass the test after these changes?

I have just locally run the tests for these packages (with both `FixedEffects.jl` and `Vcov.jl` set to be the development versions in the `Manifest.toml`). All tests passed.