@junyuan-chen 2021-02-25
There is one last thing I forgot to mention earlier. I found that `refsrev` does not handle the case when the `FixedEffect` involves any `missing`. In that case, `fe.refs` contains zero. Calling `push!(out[fe.refs[i]], i)` will result in `BoundsError`.

I did not make changes on that because we do not really construct any `FixedEffect` with vectors containing `missing` in any current usage. So, the current status is to let that `BoundsError` shows up in case anyone is trying to directly construct a `FixedEffect` with vectors containing `missing`.